refactor: home > client component 분리
