+++ components/home/featuredPost/FeaturedPost.module.scss
